+++ NEWS.md
@@ -1,5 +1,7 @@
 # mlr3learners (development version)
 
+* feat: `classif.xgboost` and `regr.xgboost` now implement `$best_valid_scores`.
+
 # mlr3learners 0.15.1
 
 * compatibility: mlr3 1.7.2